bin: fix packet size log format
(cherry picked from commit ab8a84249c28a369d639fbd69bf693cd158819f1)

+++ bin_interface.c
@@ -444,8 +444,8 @@ void call_callbacks(char* buffer, struct receive_info *rcv)
 		return;
 	}
 	if (pkg_len > BIN_MAX_BUF_LEN) {
-		LM_ERR("BIN packet size %u exceeds max size %u\n",
-				pkg_len, BIN_MAX_BUF_LEN);
+		LM_ERR("BIN packet size %u exceeds max size %zu\n",
+				pkg_len, (size_t)BIN_MAX_BUF_LEN);
 		return;
 	}
